reduce size required
I found a way to get rid of the SCT transformed data and still have the umap plots look okay. Turns out the SCT data was transformed by applying the natural log of the counts data + 1. So i simply applied that to the normalized data and removed the larger SCT datasets